eva_onboarding_concierge/cloud_run_app/requirements.txt
# Streamlit and web framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud and ADK dependencies
google-adk>=1.0.0
google-cloud-aiplatform[agent_engines]>=1.93.1
google-genai>=1.16.1
vertexai>=1.38.0

# Core dependencies
pydantic>=2.10.6,<3.0.0
python-dotenv>=1.0.1
typing-extensions>=4.0.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.7.0

# Additional utilities
requests>=2.32.3,<3.0.0
asyncio-mqtt>=0.11.0

# For health checks
curl>=0.0.1
